Unify cubic and linear path interpolators.
