+++ src/utils/d3-decorators.js
@@ -3,7 +3,6 @@ import {utilsDom} from './utils-dom';
 import {utilsDraw} from './utils-draw';
 import {default as d3} from 'd3';
 import createPathPointsInterpolator from './path/path-points-interpolator';
-import createCubicPathInterpolator from './path/cubic-path-interpolator';
 import toCubic from './path/smooth-cubic-line';
 
 var d3getComputedTextLength = () => utils.memoize(
@@ -591,10 +590,7 @@ var d3_createPathTween = (attr, pathStringBuilder, pointConvertor, idGetter, typ
         var pointsTo = type === 'cubic' ? toCubic(points) : points;
         var pointsFrom = this[pointStore];
 
-        var interpolate = {
-            linear: createPathPointsInterpolator,
-            cubic: createCubicPathInterpolator
-        }[type](pointsFrom, pointsTo);
+        var interpolate = createPathPointsInterpolator(pointsFrom, pointsTo, type);
 
         return (t) => {
             if (t === 0) {
